chore: circular dep removal
