+++ .github/workflows/dependency-review.yml
@@ -35,11 +35,11 @@ jobs:
         run: |
           npx madge --circular --ts-config tsconfig.json --extensions ts,js src/ > tmp.log || true # Force exit 0 for post-processing
           tail -n +4 tmp.log > circular-deps.log
-          if [ $(wc -l < circular-deps.log) -gt 2 ]; then
-            echo "circular-deps.log has more than 2 circular dependencies."
+          if [ $(wc -l < circular-deps.log) -gt 1 ]; then
+            echo "circular-deps.log has more than 1 circular dependency."
             wc -l circular-deps.log
             exit 1
           else
-            echo "circular-deps.log has 2 or fewer circular dependencies."
+            echo "circular-deps.log has 1 or fewer circular dependencies."
             exit 0
           fi
